Refactor chatwidget state into modules
