+++ codex-rs/tui/src/chatwidget/plugins.rs
@@ -60,6 +60,18 @@ const PLUGIN_ROW_PREFIX_WIDTH: usize = 6;
 const LOADING_ANIMATION_DELAY: Duration = Duration::from_secs(1);
 const LOADING_ANIMATION_INTERVAL: Duration = Duration::from_millis(100);
 
+#[derive(Debug, Clone, Default)]
+pub(super) struct PluginListFetchState {
+    pub(super) cache_cwd: Option<PathBuf>,
+    pub(super) in_flight_cwd: Option<PathBuf>,
+}
+
+#[derive(Debug, Clone)]
+pub(super) struct PluginInstallAuthFlowState {
+    plugin_display_name: String,
+    next_app_index: usize,
+}
+
 struct DelayedLoadingHeader {
     started_at: Instant,
     frame_requester: FrameRequester,
@@ -474,7 +486,7 @@ impl ChatWidget {
                             self.plugin_install_apps_needing_auth.len()
                         )),
                     );
-                    self.plugin_install_auth_flow = Some(super::PluginInstallAuthFlowState {
+                    self.plugin_install_auth_flow = Some(PluginInstallAuthFlowState {
                         plugin_display_name,
                         next_app_index: 0,
                     });
